requirements.txt
Flask==2.0.2
requests
python-whois